Abstracted taskbar, new page created
