+++ frontend/lib/pages/widgets/login/sign_up.dart
@@ -5,6 +5,7 @@ import 'package:frontend/widgets/snackbar.dart';
 import 'package:http/http.dart' as http;
 import 'dart:convert';
 
+
 import 'package:frontend/main.dart';
